hw5q4 code complete

+++ hw5/cs285/agents/awac_agent.py
@@ -51,7 +51,7 @@ class AWACAgent(DQNAgent):
     def get_qvals(self, critic, obs, action):
         # get q-value for a given critic, obs, and action
         qa_value = critic.q_net(obs)
-        q_value = torch.gather(qa_value, 1, action.unsqueeze(1)).squeeze(1) 
+        q_value = torch.gather(qa_value, 1, action.type(torch.int64).unsqueeze(1)).squeeze(1) 
         return q_value
 
     def estimate_advantage(self, ob_no, ac_na, re_n, next_ob_no, terminal_n, n_actions=10):
